+++ CHANGELOG.md
@@ -9,6 +9,7 @@ New:
 - Add `ValueExt::string()` as the preferred method for converting from `OsString` into `String`. Unlike [`OsString::into_string()`](https://doc.rust-lang.org/std/ffi/struct.OsString.html#method.into_string) it has a normal error type so it's compatible with catch-all error types like [`anyhow::Error`](https://docs.rs/anyhow/latest/anyhow/struct.Error.html).
   - `into_string()?` will stay supported for the time being.
 - Add `RawArgs::as_slice()` for unlimited lookahead.
+- Add `Parser::try_raw_args()` to get raw arguments without consuming a value in case of failure.
 - `Parser` implements `Clone`, `Send`, and `Sync`. Its `Debug` output now shows the remaining arguments.
 
 Changes: